docs/backend/requirements.txt
# FastAPI & Server
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Validation & Settings
pydantic==2.6.1
pydantic-settings==2.1.0

# Scraping
playwright==1.41.2

# Text Extraction
PyMuPDF==1.23.8
python-docx==1.1.0
openpyxl==3.1.2
xlrd==2.0.1

# OCR (CPU only)
paddlepaddle==2.6.0
paddleocr==2.7.3
Pillow==10.2.0

# HTTP Client
httpx==0.26.0

# Utilities
python-dateutil==2.8.2
